Parkinson_disease_Prediction/requirements.txt
Flask
scikit-learn
numpy
pandas
PyMuPDF
python-docx
spacy
rapidfuzz
